requirements.txt
streamlit
sentence-transformers
torchvision
faiss-cpu
pypdf
google-generativeai
python-dotenv
numpy
tqdm